Provider mixin and dont pass sets when context can be used
