+++ src/layouts/home-assistant.ts
@@ -11,6 +11,7 @@ import { getRecorderInfo } from "../data/recorder";
 import "../resources/custom-card-support";
 import { HassElement } from "../state/hass-element";
 import QuickBarMixin from "../state/quick-bar-mixin";
+import { RelatedContextProviderMixin } from "../state/related-context-provider-mixin";
 import type { HomeAssistant, Route } from "../types";
 import { storeState } from "../util/ha-pref-storage";
 import {
@@ -34,7 +35,9 @@ const panelUrl = (path: string) => {
 };
 
 @customElement("home-assistant")
-export class HomeAssistantAppEl extends QuickBarMixin(HassElement) {
+export class HomeAssistantAppEl extends RelatedContextProviderMixin(
+  QuickBarMixin(HassElement)
+) {
   @state() private _route: Route;
 
   @state() private _databaseMigration?: boolean;
